+++ Sources/KeyboardKit/Callouts/CalloutCurve.swift
@@ -10,12 +10,12 @@ import CoreGraphics
 import SwiftUI
 
 /**
- [DEPRECATED] This will be made internal in KeyboardKit 8.0.
+ This will be made internal in KeyboardKit 8.0.
  
  This curve can be used to smoothen the part where a callout
  bubble meets the button area.
  */
-public struct CalloutCurve: Shape {
+struct CalloutCurve: Shape {
     
     public func path(in rect: CGRect) -> Path {
         var path = Path()